Drop asterisks in non-template presets
Distributions, extras, and traits are now named without asterisks
marking up-to-date status. This is a slight reduction in clarity
since previously there was an implicit outdated/updated based on
the presence of the asterisk. An alternative would be to add (Outdated)
for a clear indicator but it felt a bit too wide and would require
future renaming.