Adds support for template data `outdated` key
This key is then used in the navbar to replace the current manual
Outdated renaming of templates. Note that the buildTemplateOutdated
translation key is simply concatenated to the build template name which
may not be suitable for non-english languages. However, a brief google
translate experiment makes it seem like most languages should work.