Add D/D Alac Specter; Remove one of the Sc/D alac spectre variants
Sc/D Alac is very bad but of potential historical interest (and can be
updated if it's ever less bad).

Adds a new double-sigil since d/d alac specter can proc doom + geo on
cooldown with shroud and frequent weapon swapping. As a result, the
extras template needs to not have doom/geo/this new sigil on the main
hand because it will think that geo (or doom) + the new geo-doom sigil
is super powerful.