June 25 Balance (#764)
* Mark Ele templates as outdated

- Alac temp: just generally outdated
- CWeaver Scepter: scepter skills are changing

* Add low health Flow like Water; Unsplit high health version

"Flow like Water: Increased the bonus strike damage while above the
health threshold from 5% to 10% in PvP and WvW. This trait now also
grants 5% bonus strike damage while under the health threshold. Reduced
the health threshold from 75% to 50%. Reduced the internal cooldown from
10 seconds to 4 seconds."

* Add Soothing Power trait

"Soothing Power: This trait now also grants 300 vitality to the
elementalist."

* Merge Gathered Focus versions

"Gathered Focus: The additional concentration from this trait no longer
has a health-threshold requirement."

* Mark PHolo as outdated

nerfed :(

* Mark CVirt as outdated

* Add damage modifier to Infinite Forge

10% on blade skill damage, sensible default to be added when PVirt is
updated

Main reason for representing this as a proper trait instead of baking it
into the distribution is that I think it would be cool to approximate
a comparison between infinite forge and bloodsong but I could easily be
convinced to not do the extra work.

* Add Heightened Foci

An outrageous Warrior trait that requires a bit of kludging to reflect
the different adrenaline level effects in a way that hopefully makes the
most sense. Really should get around to adding a heal warrior template
one of these days.

* Add All for One's damage modifier

* Update disclaimer

* Mark CWB as outdated

Rushing Justice buffed

* Default input of new traits to ???

* Mark Firebrand templates as outdated

Oh so *that's* the mantra charging change

* Mark Soothing Power as unconverted

* Mark Condi Druid as outdated