Add C Q Cata P/D template; Reduce Condi Boon to 0 required bd
Received permission from Maragon to use their log

Reduces condi boon's boon duration to 0 so that it can be used as a
generic condi boon affix instead of what I think was legacy renegade
alac. It was completely unused so this represents a good opportunity to
consolidate since there's a lot of builds that can default to 0 bd.